refactor(server): use fastify-static for assets only
This enables using `/*` to return `index.html`for any path not matched
by Fastify, which is needed for HTML5 routing.

Before, it was going to be required to do every base path for the UI
such as `/show` or `/movie`. Now the server does not need to be aware
of the UI routes.