fix(all): use HTTPS for client instead of HTTP
As of Chrome 72, Chromecast web sender apps must be running on HTTPS
[1].

Since TV uses private IPs only, we cannot use trusted certs, so we must
create self-signed certs. This makes Chromecast happy from the web
sender app happy.

Unfortunately, the actual Chromecast device does not seem to like
self-signed certs. So, we're still using an HTTP server to link to the
videos directly.

We cannot simply use this HTTP server to retrieve our list of shows in
the webapp because AJAX calls to HTTP servers from an HTTPS source are
blocked. So, we're using a proxy that calls the HTTP server behind the
scenes. But the Chromecast device only ever gets the direct HTTP link.